+++ libs/@guardian/react-crossword/src/components/Layout.tsx
@@ -2,28 +2,11 @@ import { css } from '@emotion/react';
 import { headlineBold17, space } from '@guardian/source/foundations';
 import { textSans12, textSans14 } from '@guardian/source/foundations';
 import type { ReactNode } from 'react';
-import { useMemo } from 'react';
 import { memo } from 'react';
 import type { Direction } from '../@types/Direction';
-import { useData } from '../context/Data';
 import { useTheme } from '../context/Theme';
-
-export const useGridWidth = () => {
-	const { gutter, cellSize } = useTheme();
-	const { dimensions } = useData();
-
-	return useMemo(
-		() => Math.max((cellSize + gutter) * dimensions.cols + gutter, 300),
-		[cellSize, gutter, dimensions.cols],
-	);
-};
-
-export const useWidthForCols = (cols: number) => {
-	const gridWidth = useGridWidth();
-	const { clueMinWidth } = useTheme();
-
-	return gridWidth + clueMinWidth * cols + 'px';
-};
+import { useGridWidth } from '../hooks/useGridWidth';
+import { useWidthForCols } from '../hooks/useWidthForCols';
 
 export const Wrapper = memo(({ children }: { children: ReactNode }) => {
 	const { text, clueMaxWidth } = useTheme();
