+++ libs/@guardian/react-crossword/src/hooks/useWidthForCols.ts
@@ -1,0 +1,9 @@
+import { useTheme } from '../context/Theme';
+import { useGridWidth } from './useGridWidth';
+
+export const useWidthForCols = (cols: number) => {
+	const gridWidth = useGridWidth();
+	const { clueMinWidth } = useTheme();
+
+	return gridWidth + clueMinWidth * cols + 'px';
+};
